Añado step para ejecutar las migraciones y configuro los deploys

+++ app/controllers/reset_controller.rb
@@ -1,5 +1,9 @@
 DeliveryApi::App.controllers do
   post '/reset' do
+    if ENV['RACK_ENV'] == 'prod'
+      status 401
+      return 'ERROR'
+    end
     ClientRepository.new.delete_all
     UserRepository.new.delete_all
     'Ok'
